+++ components/common/pagination/index.tsx
@@ -9,6 +9,7 @@ import {
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import PageNumber from './PageNumber'
 import useDeliveryShippingList from 'pages/api/query/useDeliveryShippingList'
+import useIsMobile from '../custom/isMobile'
 
 export default function Pagination(props: {
   page: number
@@ -26,23 +27,29 @@ export default function Pagination(props: {
     }
   }, [data, props, status])
 
+  const isMobile = useIsMobile()
+
   let arr = []
-  let pageCount = 10
+  let pageCount = isMobile ? 5 : 10
 
   const handleClickPageButton = (event: MouseEvent<HTMLButtonElement>) => {
     const { name, innerText } = event.currentTarget
 
     if (name == 'prevPage') {
-      props.setPage(props.page - 1 - ((props.page - 1) % 10))
-      props.setStartPage(props.page - 10 - ((props.page - 1) % 10))
+      props.setPage(props.page - 1 - ((props.page - 1) % pageCount))
+      props.setStartPage(
+        props.page - pageCount - ((props.page - 1) % pageCount)
+      )
     } else if (name == 'nextPage') {
-      props.setStartPage(props.page + 10 - ((props.page - 1) % 10))
-      props.setPage(props.page + 10 - ((props.page - 1) % 10))
+      props.setStartPage(
+        props.page + pageCount - ((props.page - 1) % pageCount)
+      )
+      props.setPage(props.page + pageCount - ((props.page - 1) % pageCount))
     } else if (name == 'firstPage') {
       props.setStartPage(1)
       props.setPage(1)
     } else if (name == 'lastPage') {
-      props.setStartPage(props.totalPage - (props.totalPage % 10) + 1)
+      props.setStartPage(props.totalPage - (props.totalPage % pageCount) + 1)
       props.setPage(props.totalPage)
     } else {
       props.setPage(Number(innerText.toString()))
@@ -73,33 +80,35 @@ export default function Pagination(props: {
 
   return (
     <div className="flex justify-center flex-row items-center">
-      <button
-        type="button"
-        name="firstPage"
-        title="첫 페이지로 이동"
-        onClick={(event) => handleClickPageButton(event)}
-        disabled={props.page == 1}
-        className={`${props.page == 1 ? 'fa-disabled' : ''}`}
-      >
-        <FontAwesomeIcon
-          icon={faAngleDoubleLeft}
-          width="15px"
-          height="15px"
-          style={{
-            position: 'relative',
-            margin: '0px 10px',
-            cursor: 'pointer',
-          }}
-        />
-      </button>
+      {!isMobile && (
+        <button
+          type="button"
+          name="firstPage"
+          title="첫 페이지로 이동"
+          onClick={(event) => handleClickPageButton(event)}
+          disabled={props.page == 1}
+          className={`${props.page == 1 ? 'fa-disabled' : ''}`}
+        >
+          <FontAwesomeIcon
+            icon={faAngleDoubleLeft}
+            width="15px"
+            height="15px"
+            style={{
+              position: 'relative',
+              margin: '0px 10px',
+              cursor: 'pointer',
+            }}
+          />
+        </button>
+      )}
       <button
         type="button"
         name="prevPage"
         title="이전 페이지 그룹으로 이동"
         onClick={(event) => handleClickPageButton(event)}
-        disabled={Math.floor((props.page - 1) / 10) == 0}
+        disabled={Math.floor((props.page - 1) / pageCount) == 0}
         className={`${
-          Math.floor((props.page - 1) / 10) == 0 ? 'fa-disabled' : ''
+          Math.floor((props.page - 1) / pageCount) == 0 ? 'fa-disabled' : ''
         }`}
       >
         <FontAwesomeIcon
@@ -126,9 +135,13 @@ export default function Pagination(props: {
         name="nextPage"
         title="다음 페이지 그룹으로 이동"
         onClick={(event) => handleClickPageButton(event)}
-        disabled={props.totalPage - props.startPage <= 9 ? true : false}
+        disabled={
+          props.totalPage - props.startPage <= pageCount - 1 ? true : false
+        }
         className={`${
-          props.totalPage - props.startPage <= 9 ? 'fa-disabled' : ''
+          props.totalPage - props.startPage <= pageCount - 1
+            ? 'fa-disabled'
+            : ''
         }`}
       >
         <FontAwesomeIcon
@@ -142,25 +155,27 @@ export default function Pagination(props: {
           }}
         />
       </button>
-      <button
-        type="button"
-        name="lastPage"
-        title="마지막 페이지로 이동"
-        onClick={(event) => handleClickPageButton(event)}
-        disabled={props.page == props.totalPage}
-        className={`${props.page == props.totalPage ? 'fa-disabled' : ''}`}
-      >
-        <FontAwesomeIcon
-          icon={faAngleDoubleRight}
-          width="15px"
-          height="15px"
-          style={{
-            position: 'relative',
-            margin: '0px 10px',
-            cursor: 'pointer',
-          }}
-        />
-      </button>
+      {!isMobile && (
+        <button
+          type="button"
+          name="lastPage"
+          title="마지막 페이지로 이동"
+          onClick={(event) => handleClickPageButton(event)}
+          disabled={props.page == props.totalPage}
+          className={`${props.page == props.totalPage ? 'fa-disabled' : ''}`}
+        >
+          <FontAwesomeIcon
+            icon={faAngleDoubleRight}
+            width="15px"
+            height="15px"
+            style={{
+              position: 'relative',
+              margin: '0px 10px',
+              cursor: 'pointer',
+            }}
+          />
+        </button>
+      )}
 
       <style jsx>{`
         .fa-disabled {
