+++ pages/mobile/mypage/delivery.tsx
@@ -1,11 +1,8 @@
 import LayoutMobile from '@components/LayoutMobile'
+import ContentsDelivery from '@components/mypage/ContentsDelivery'
 
 export default function Delivery() {
-  return (
-    <div className="flex justify-center h-10">
-      <strong>작성 중인 페이지입니다.</strong>
-    </div>
-  )
+  return <ContentsDelivery />
 }
 Delivery.getLayout = function getLayout(page: React.ReactElement) {
   return <LayoutMobile>{page}</LayoutMobile>
